Back end

+++ backend/controllers/contactController.js
@@ -2,14 +2,14 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel.js");
 
 // Get all the contacts
-// @acess public
+// @acess private
 // route -> GET /api/contacts
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 // Create New  contacts
-// @acess public
+// @acess private
 // route -> POST /api/contacts
 const createContact = asyncHandler(async (req, res) => {
   console.log(req.body);
@@ -22,11 +22,12 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 // GET single contact
-// @acess public
+// @acess private
 // route -> GET /api/contacts/:id
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -38,13 +39,18 @@ const getContact = asyncHandler(async (req, res) => {
   res.status(200).json(contact);
 });
 // Update single contact
-// @acess public
+// @acess private
 // route -> PUT /api/contacts/:id
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact couldn't be found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permission to update the contact");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -56,13 +62,17 @@ const updateContact = asyncHandler(async (req, res) => {
   res.status(200).json(contact);
 });
 // Delete single contact
-// @acess public
+// @acess private
 // route -> DEL /api/contacts/:id
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact couldn't be found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permission to Delete the contact");
   }
   const updatedContact = await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
